+++ ARTG170/Assets/GameNameTBD/Scripts/Managers/SceneChangeManager.cs
@@ -1,21 +1,71 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
-public class SceneChangeManager : SingletonBase<SceneChangeManager>
+// https://www.youtube.com/watch?v=3I5d2rUJ0pE&ab_channel=CodeMonkey
+public class SceneChangeManager : MonoBehaviour
 {
-
-
     [Header("Debug")]
     [SerializeField]
     Logger _logger;
 
+    private class LoadingMonoBehavior: MonoBehaviour { }
+    public enum Scene {
+        MainGameScene,
+        Loading,
+        MainMenuScene
+    }
+
+    private static Action onLoaderCallback;
+    private static AsyncOperation loadingAsyncOperation;
+
+    public static void Load(Scene scene)
+    {
+        onLoaderCallback = () =>
+        {
+            GameObject loadingGameObject = new GameObject("Loading GameObject");
+            loadingGameObject.AddComponent<LoadingMonoBehavior>().StartCoroutine(LoadSceneAsync(scene));
+        };
+
+        SceneManager.LoadScene(Scene.Loading.ToString());
+    }
+
+    private static IEnumerator LoadSceneAsync(Scene scene)
+    {
+        yield return null;
+        loadingAsyncOperation = SceneManager.LoadSceneAsync(scene.ToString());
+        while (!loadingAsyncOperation.isDone)
+        {
+            yield return null;
+        }
+    }
+
+    public static float GetLoadingProgress()
+    {
+        if (loadingAsyncOperation != null)
+        {
+            return loadingAsyncOperation.progress;
+        } 
+        else
+        {
+            return 1f;
+        }
+    }
+
+    public static void LoaderCallback()
+    {
+        if (onLoaderCallback != null)
+        {
+            onLoaderCallback();
+            onLoaderCallback = null;
+        }
+    }
 
 
     private void Start()
     {
         _logger.Log("Loading SceneManager!");
-        
     }
 }