+++ ARTG170/Assets/GameNameTBD/Scripts/UI/HUDMenu.cs
@@ -4,5 +4,17 @@ using UnityEngine;
 
 public class HUDMenu : MenuManager
 {
-
+    [SerializeField] private UIManager _uiManager;
+    protected override void InnerAwake()
+    {
+        menuType = GameMenu.GameHUD;
+    }
+    private void Update()
+    {
+        // On User Escape, we go to the settings.
+        if (Input.GetKeyDown(KeyCode.Escape) && _uiManager.getCurrentMenu() == GameMenu.GameHUD)
+        {
+            _uiManager.GoToMenu(GameMenu.Settings);
+        }
+    }
 }